pyproject.toml
[project]
name = "bot-database-checking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "faker==37.0.2",
    "pillow==11.1.0",
    "python-dotenv==1.0.1",
    "python-telegram-bot==22.0",
    "sqlmodel==0.0.24",
]
